Plotter/, review

+++ software/bench_scripts/classes/Plotter/EachPlotter.php
@@ -1,18 +1,16 @@
 <?php
 namespace Plotter;
 
-final class EachPlotter implements IPlotter
+final class EachPlotter implements IGroupPlotter
 {
 
     private const template = __DIR__ . '/templates/each.plot.php';
 
     private \Plot $plot;
 
-    private string $csvPath;
+    private string $dirName;
 
-    private string $fileName;
-
-    private string $basePath;
+    private string $queryName;
 
     private array $data;
 
@@ -31,13 +29,13 @@ final class EachPlotter implements IPlotter
         return \Plot::PROCESS_EACH;
     }
 
-    public function plot(array $csvPaths): void
+    public function plot(string $dirName, array $queriesName): void
     {
+        \wdPush($dirName);
         $plot = $this->plot;
-        $this->csvPath = $csvPath = $csvPaths[0];
-        $this->fileName = \baseName($csvPath, ".csv");
-        $this->basePath = \dirname($csvPath);
-        $this->data = $this->plot->getData()[$this->csvPath];
+        $this->dirName = $dirName;
+        $this->queryName = $queriesName[0];
+        $this->data = $this->plot->getData()["$dirName/{$this->queryName}"];
 
         $this->writeDat();
 
@@ -45,20 +43,24 @@ final class EachPlotter implements IPlotter
             'PLOT' => $this->plot,
             'PLOTTER' => $this
         ]);
+
         $plotFileName = $this->getFileName('.plot');
-        \file_put_contents($plotFileName, $contents);
+        if (! \is_file($plotFileName))
+            \file_put_contents($plotFileName, $contents);
 
         $outFileName = $this->getFileName('.png');
-        $cmd = "gnuplot '$plotFileName' > '$outFileName'";
-        echo "plotting $outFileName\n";
-
-        system($cmd);
+        if (! \is_file($outFileName)) {
+            $cmd = "gnuplot '$plotFileName' > '$outFileName'";
+            echo "plotting $outFileName\n";
+            system($cmd);
+        }
+        \wdPop();
     }
 
     // ========================================================================
     public function getFileName(string $suffix = ''): string
     {
-        return \Plot::plotterFileName($this, $this->csvPath, $suffix);
+        return \Plot::plotterFileName($this, $this->queryName, $suffix);
     }
 
     public function getData(): array
@@ -71,7 +73,8 @@ final class EachPlotter implements IPlotter
     {
         echo "measure real cpu\n";
 
-        foreach (\array_filter($data, 'Plot::isTimeMeasure') as $group => $measure) {
+        foreach (\array_filter($data, '\Measures::isTimeMeasure') as $group => $measure) {
+            $measure = \Measures::decodeTimeMeasure($measure);
             $r = max(0, $measure['r']);
             $c = max(0, $measure['c']);
             $realRemain = max(0, $r - $c);
@@ -83,6 +86,10 @@ final class EachPlotter implements IPlotter
     private function writeDat()
     {
         $file = $this->getFileName('.dat');
+
+        if (\is_file($file))
+            return;
+
         echo "Writing $file\n";
         $content = \get_ob(fn () => $this->echoDat($this->data));
